♻ split parser into seperate files
